fix for make dependency bug
Apparently there are two identical rules to build receivers.cpp with stub_gen,
but only one of them depended on stub_gen actually being built.  This only is
manifested for uw builds with a large -jN number for make (our el5 koji server
has -j24).  This patch adds the dependency to the cmake rules, but whether it
a problem that there are two separate identical rules to build receivers.cpp
is another question...



git-svn-id: file:///p/condor/workspaces/vdt/svn/native/redhat/upcoming@17146 4e558342-562e-0410-864c-e07659590f8c