cleanup: reorder & group patches more sensibly (SOFTWARE-2161)
well, at least to me, at the moment.

a handful of the patches get applied to the hadoop-bigtop-packaging
sources and the rest to the main hadoop sources; and we pushd into the
bigtop-packaging dir to apply the first set.  let's make it easier to
add more if we need to and still apply them in an increasing order.

git-svn-id: file:///p/condor/workspaces/vdt/svn/native/redhat/branches/software-2161@22541 4e558342-562e-0410-864c-e07659590f8c